Refactor NewSuggestion to ISuggestion
