+++ src/pages/suggestions/new/index.tsx
@@ -8,7 +8,7 @@ import { Area } from "@/constants/area"
 import { useSelector } from "react-redux"
 import { AppState } from "@/redux/store"
 import { FormProvider, useForm } from "react-hook-form"
-import { InputsSongInformation } from "@/interfaces/new-suggestion"
+import { InputsSongInformation } from "@/interfaces/suggestion"
 import InstrumentsArea from "@/components/new-suggestion/areas/instruments/instruments.area"
 import { Database } from "@/types/database"
 import { Instrument } from "@/types/database-types"
